Add tests, improve on given comments and fix issue with I1
